refactor: split ui session planner
