+++ src/ui/session_planner.rs
@@ -1,0 +1,201 @@
+use super::*;
+
+pub(super) fn render_session_planner(frame: &mut Frame, app: &App) {
+    let area = frame.area();
+    let outer = centered_rect(52, 70, area);
+
+    let block = Block::default()
+        .borders(Borders::ALL)
+        .title(" Session Planner ")
+        .title_alignment(Alignment::Center)
+        .style(Style::default().fg(Color::Cyan));
+    frame.render_widget(block, outer);
+
+    let inner = Layout::default()
+        .direction(Direction::Vertical)
+        .margin(2)
+        .constraints([
+            Constraint::Length(1), // current task
+            Constraint::Min(4),    // task labels list
+            Constraint::Length(3), // task label input
+            Constraint::Length(1), // feedback
+            Constraint::Length(3), // hints
+        ])
+        .split(outer);
+
+    render_session_planner_selected_task(frame, app, inner[0]);
+    render_session_planner_labels(frame, app, inner[1]);
+    render_session_planner_input(frame, app, inner[2]);
+    render_session_planner_feedback(frame, app, inner[3]);
+    render_session_planner_hints(frame, app, inner[4]);
+}
+
+fn render_session_planner_selected_task(frame: &mut Frame, app: &App, area: Rect) {
+    let selected_text = app.selected_task_label.as_ref().map_or_else(
+        || "Selected task: none (required before focus starts)".to_string(),
+        |label| format!("Selected task: {label}"),
+    );
+    frame.render_widget(
+        Paragraph::new(Line::from(selected_text))
+            .style(Style::default().fg(Color::White))
+            .wrap(Wrap { trim: true }),
+        area,
+    );
+}
+
+fn render_session_planner_labels(frame: &mut Frame, app: &App, area: Rect) {
+    if app.task_labels.is_empty() {
+        frame.render_widget(
+            Paragraph::new("No task labels yet. Press [a] to add one.")
+                .style(Style::default().fg(Color::DarkGray))
+                .block(
+                    Block::default()
+                        .borders(Borders::ALL)
+                        .title(" Task Labels "),
+                ),
+            area,
+        );
+        return;
+    }
+
+    let display_labels = app.planner_labels_for_display();
+    let items: Vec<ListItem> = display_labels
+        .iter()
+        .map(|label| {
+            let marker = if app
+                .selected_task_label
+                .as_ref()
+                .is_some_and(|selected| selected.eq_ignore_ascii_case(label))
+            {
+                "✓"
+            } else {
+                " "
+            };
+            let mut badges = Vec::new();
+            if app.is_task_label_favorite(label) {
+                badges.push("★");
+            }
+            if app.is_task_label_archived(label) {
+                badges.push("archived");
+            }
+            let suffix = if badges.is_empty() {
+                String::new()
+            } else {
+                format!(" ({})", badges.join(", "))
+            };
+            ListItem::new(format!(" {marker} {label}{suffix}"))
+        })
+        .collect();
+    let list = List::new(items)
+        .block(
+            Block::default()
+                .borders(Borders::ALL)
+                .title(" Task Labels "),
+        )
+        .highlight_style(
+            Style::default()
+                .fg(Color::Black)
+                .bg(Color::White)
+                .add_modifier(Modifier::BOLD),
+        )
+        .highlight_symbol("▶ ");
+    let mut state = ListState::default();
+    state.select(Some(
+        app.planner_selection_index
+            .min(display_labels.len().saturating_sub(1)),
+    ));
+    frame.render_stateful_widget(list, area, &mut state);
+}
+
+fn render_session_planner_input(frame: &mut Frame, app: &App, area: Rect) {
+    let input_title = if app.planner_input_active {
+        match app.planner_input_mode {
+            Some(PlannerInputMode::Rename) => " Rename task label ",
+            _ => " Add task label ",
+        }
+    } else {
+        " Task label input ([a] add / [e] rename / [f] favorite / [x] archive) "
+    };
+    let input_text = if app.planner_input_active {
+        format!("{}|", app.planner_input)
+    } else {
+        "Use [a] add, [e] rename, [f] favorite, [x] archive, [d/Del] delete highlighted".to_string()
+    };
+    frame.render_widget(
+        Paragraph::new(input_text)
+            .style(if app.planner_input_active {
+                Style::default().fg(Color::White)
+            } else {
+                Style::default().fg(Color::DarkGray)
+            })
+            .block(Block::default().borders(Borders::ALL).title(input_title)),
+        area,
+    );
+}
+
+fn render_session_planner_feedback(frame: &mut Frame, app: &App, area: Rect) {
+    if let Some(feedback) = app.planner_feedback.as_ref() {
+        let (prefix, color) = match feedback.level {
+            PlannerFeedbackLevel::Success => ("✓", Color::Green),
+            PlannerFeedbackLevel::Warning => ("⚠", Color::Yellow),
+        };
+        frame.render_widget(
+            Paragraph::new(format!("{prefix}  {}", feedback.message))
+                .style(Style::default().fg(color))
+                .alignment(Alignment::Center)
+                .wrap(Wrap { trim: true }),
+            area,
+        );
+    }
+}
+
+fn planner_recent_quick_pick_text(app: &App) -> String {
+    let recent = app.planner_recent_labels(PLANNER_RECENT_LABEL_LIMIT);
+    if recent.is_empty() {
+        return "Recent: none yet".to_string();
+    }
+
+    let mut parts = Vec::new();
+    for (index, label) in recent.iter().enumerate() {
+        if index == 0 {
+            parts.push(format!("[r/1] {label}"));
+        } else {
+            parts.push(format!("[{}] {label}", index + 1));
+        }
+    }
+    format!("Recent: {}", parts.join("  "))
+}
+
+fn render_session_planner_hints(frame: &mut Frame, app: &App, area: Rect) {
+    let hints = if app.planner_input_active {
+        let commit_line = match app.planner_input_mode {
+            Some(PlannerInputMode::Rename) => "Input: rename label, then [Enter]",
+            _ => "Input: type task label, then [Enter]",
+        };
+        vec![
+            Line::from(commit_line),
+            Line::from("Input: [Esc] Cancel"),
+            Line::from(if app.strict_mode_enforced_for_focus() {
+                "View: [q/Ctrl-C] Quit (Locked)"
+            } else {
+                "View: [q/Ctrl-C] Quit"
+            }),
+        ]
+    } else {
+        vec![
+            Line::from(
+                "Planner: [↑/↓] Move  [Enter] Select  [a] Add  [e] Rename  [f] Favorite  [x] Archive  [d/Del] Delete",
+            ),
+            Line::from(format!(
+                "{}  |  Archived labels stay visible and cannot be selected",
+                planner_recent_quick_pick_text(app)
+            )),
+            Line::from(if app.strict_mode_enforced_for_focus() {
+                "View: [t/Esc] Back  [q/Ctrl-C] Quit (Locked)"
+            } else {
+                "View: [t/Esc] Back  [q/Ctrl-C] Quit"
+            }),
+        ]
+    };
+    render_hint_lines(frame, area, hints);
+}
